Use usize for the counter value

+++ mac-n-cheese/compiler/src/sieve_compiler/circuit_ir.rs
@@ -138,7 +138,7 @@ pub struct CounterInfo {
     pub num_env: u64,
     pub field_type: FieldType,
     pub num_wires: u64,
-    pub value: u64,
+    pub value: usize,
 }
 
 #[derive(Debug)]
